feat: isFullyOverlapped 함수 분리
